requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
deepface==0.0.93
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mtcnn==1.0.0
namex==0.0.9
numpy==2.1.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spotipy==2.25.1
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tf_keras==2.19.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
